Create countFaults.java
Implement a prototype ervice to automate the detection and replacement of faulty servers to improve the availability of and application.
There are n server with ids s1,s2,…,sn and an array of strings, logs, of size m. Log format is “<server_id><success/error> “, the id of the server, and the status of the processed request. If a particular	server id logs an error for three coonosecuritve request, it is considered faulty and is replaced with a new server with the same id.

Given n and the array logs, find the number oof times a faulty server was replaced.